Updated - getDispensersByIdentifier - getDispensesByIdentifier - getIssuancesByIdentifier - getDestructionsByIdentifier - getSendsByIdentifier
All now return the specific necessary fields and now properly return
data from subassets.

Note: Need to discuss if we are trying to recreate exact queries from
xchain or give the necessary information.  Right now some datatypes
returned are ints when xchain returns strings. Also, some properies arent
returnd at all because they arent present in the database.

I think the correct decision here is to return the necessary data, not
make a 1 for 1 copy.  This means freewallet wont fully function, but
we should be able to pass users tx's to sign to bridge the gap